refactor(icon): extract icon into its own component
