+++ src/Checkbox/index.js
@@ -1,42 +1,12 @@
-import React from 'react'
+import React, { Fragment } from 'react'
 import propTypes from 'prop-types'
 import cx from 'classnames'
-
-import { colors, theme } from '../theme.js'
-import styles from './styles.js'
-
-import { Indeterminate, Checked, Unchecked } from '../icons/Checkbox.js'
-
 import css from 'styled-jsx/css'
 
-const icons = css.resolve`
-    svg {
-        height: 24px;
-        width: 24px;
-        fill: ${theme.default};
-    }
-
-    .checked {
-        fill: ${colors.teal400};
-    }
-
-    .disabled {
-        fill: ${theme.disabled};
-    }
-
-    .error {
-        fill: ${theme.error};
-    }
-
-    .valid {
-        fill: ${theme.valid};
-    }
-
-    .warning {
-        fill: ${theme.warning};
-
-    }
-`
+import { colors, theme } from '../theme'
+import { Indeterminate, Checked, Unchecked } from '../icons/Checkbox'
+import { Icon } from './Icon'
+import styles from './styles'
 
 class Checkbox extends React.Component {
     state = {
@@ -67,22 +37,6 @@ class Checkbox extends React.Component {
             checked = this.state.checked,
         } = this.props
 
-        const classes = cx(icons.className, {
-            checked: checked && !valid && !error && !warning,
-            disabled,
-            valid,
-            error,
-            warning,
-        })
-
-        let icon = <Unchecked className={classes} />
-
-        if (this.state.indeterminate) {
-            icon = <Indeterminate className={classes} />
-        } else if (checked) {
-            icon = <Checked className={classes} />
-        }
-
         return (
             <label
                 className={cx('base', className, {
@@ -95,13 +49,20 @@ class Checkbox extends React.Component {
                     checked={checked}
                     disabled={this.props.disabled}
                 />
-                {icon}
+
+                <Icon
+                    checked={checked}
+                    disabled={disabled}
+                    valid={valid}
+                    error={error}
+                    warning={warning}
+                    indeterminate={this.state.indeterminate}
+                />
 
                 <span className={cx('label', { required })}>
                     {this.props.label}
                 </span>
 
-                {icons.styles}
                 <style jsx>{styles}</style>
             </label>
         )
